Add preserveEntrypoint option to splitChunks cacheGroups
I am working in a complicated app that has a bunch of bundles on every
page. Most of these are "global" bundles and appear on every page. So, I
have configured webpack to prevent any modules that appear in these
bundles from appearing in any other bundle via splitChunks. More context
on this here:

  https://stackoverflow.com/questions/49163684/how-to-configure-webpack-4-to-prevent-chunks-from-list-of-entry-points-appearing

Unfortunately, it seems that when I configure splitChunks in this way,
it causes the global bundles to not be executed. In the debugging I've
done, it seems that this splitChunks configuration causes the
executeModules part of the bundle that the runtime chunk is looking for
to be undefined.

I believe this happens in the SplitChunks plugin to any reused chunk
that was already an entry point. I'm not entirely sure why this was done
this way, so to address my use-case, I am adding an option to
cacheGroups that allows the entrypoint to be preserved when moving
chunks into it.

Fixes #7230